Fixed sonar and wall crash bugs
The sonar bug resulted from displaying results from the previous scan in
the new robot frame. I believe it was just a display problem.

The program no longer crashes when the robot hits a wall. The problem
before was that if a particle ended up inside an obstacle the likelihood
gave a nan which crashed the resampling procedure. Now such particles
are thrown out by giving them zero weight. I think this is a reasonable
thing to do.